fix(ci): resolve mypy errors and add missing dependencies
- Add scikit-learn and scipy to dev dependencies in pyproject.toml
- Fix mypy type errors in metrics.py, labeling.py, golden_dataset.py, ablation.py
- Fix flaky pair generation test with exhaustive iteration
- Update wiki milestones to reflect M0-M12 status
- Remove emojis from wiki per project guidelines

+++ python/sentinel_ml/evaluation/golden_dataset.py
@@ -359,7 +359,7 @@ class RegressionResult:
         misclassified = len(self.misclassified_records)
         if total == 0:
             return 0.0
-        return (total - misclassified) / total
+        return float((total - misclassified) / total)
 
     def to_dict(self) -> dict[str, Any]:
         """Convert to dictionary for serialization."""
